Refactor Select File into its own Component
